+++ kotlin/app/src/main/java/com/gamesage/kotlin/data/local/platform/PlatformDao.kt
@@ -25,5 +25,5 @@ interface PlatformDao {
     fun observeAll(): Flow<List<PlatformEntity>>
 
     @Query("SELECT * FROM platforms WHERE id = :id")
-    suspend fun readPlatformById(id: Int): PlatformEntity?
+    suspend fun readPlatformById(id: Long): PlatformEntity?
 }
